Feat(jira): Move common issue tracker files into the issuetracker project.
