+++ build.gradle
@@ -79,6 +79,7 @@ repositories {
 dependencies {
     implementation project(':alert-common')
     implementation project(':alert-database')
+    implementation project(':alert-issuetracker')
 
     implementation 'com.synopsys.integration:polaris-common:0.12.0'
     implementation 'com.synopsys.integration:int-jira-common:0.6.1'
