+++ settings.gradle
@@ -1,4 +1,6 @@
-include 'alert-database'
 rootProject.name = 'blackduck-alert'
 include 'alert-common'
+include 'alert-database'
+include 'alert-issuetracker'
+
